+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M31AFBBFJ92AGQQNAMBGJYV5"),
+  id: z.string().default("01M31DF1VZT8EG4KG1HPD4BJCH"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>